+++ __tests__/types/__typescript__/componentSendTextBox.tsx
@@ -1,0 +1,7 @@
+import { Components } from '../../../packages/bundle';
+
+function main() {
+  const { SendTextBox } = Components;
+
+  return <SendTextBox />;
+}