+++ __tests__/types/__typescript__/dirInComposer.tsx
@@ -1,0 +1,7 @@
+import { Components } from '../../../packages/component';
+
+const { Composer } = Components;
+
+function main() {
+  return <Composer dir="rtl" />;
+}