+++ __tests__/types/__typescript__/directLineInComposer.tsx
@@ -1,0 +1,9 @@
+import { Components, createDirectLine } from '../../../packages/component';
+
+const { Composer } = Components;
+
+function main() {
+  const directLine = createDirectLine({ token: 'faketoken' });
+
+  return <Composer directLine={directLine} />;
+}