+++ __tests__/types/__typescript__/directLineObject.tsx
@@ -1,0 +1,7 @@
+import ReactWebChat, { createDirectLine } from '../../../packages/bundle';
+
+function main() {
+  const directLine = createDirectLine({ token: 'faketoken' });
+
+  return <ReactWebChat directLine={directLine} />;
+}