+++ __tests__/types/__typescript__/storeInComposer.tsx
@@ -1,0 +1,10 @@
+import { Components } from '../../../packages/component';
+import { createStore } from '../../../packages/core';
+
+const { Composer } = Components;
+
+function main() {
+  const store = createStore({});
+
+  return <Composer store={store} />;
+}