+++ __tests__/types/__typescript__/storeObject.tsx
@@ -1,0 +1,8 @@
+import ReactWebChat from '../../../packages/bundle';
+import { createStore } from '../../../packages/core';
+
+function main() {
+  const store = createStore({});
+
+  return <ReactWebChat store={store} />;
+}