+++ packages/api/src/hooks/useCreateScrollToEndButtonRenderer.ts
@@ -1,0 +1,6 @@
+import { ScrollToEndButtonCreator } from '../types/ScrollToEndButtonMiddleware';
+import useWebChatAPIContext from './internal/useWebChatAPIContext';
+
+export default function useCreateScrollToEndButtonRenderer(): ScrollToEndButtonCreator {
+  return useWebChatAPIContext().scrollToEndButtonRenderer;
+}