+++ packages/api/src/hooks/useTimeoutForSend.ts
@@ -1,0 +1,8 @@
+import DirectLineActivity from '../types/external/DirectLineActivity';
+import useStyleOptions from './useStyleOptions';
+
+export default function useTimeoutForSend(): [number | ((activity: DirectLineActivity) => number)] {
+  const [{ sendTimeout }] = useStyleOptions();
+
+  return [sendTimeout];
+}