+++ packages/api/src/index.tsx
@@ -1,4 +1,15 @@
+export * as hooks from './hooks';
+
+import ActivityMiddleware from './types/ActivityMiddleware';
+import ActivityStatusMiddleware from './types/ActivityStatusMiddleware';
+import AttachmentForScreenReaderMiddleware from './types/AttachmentForScreenReaderMiddleware';
+import AttachmentMiddleware from './types/AttachmentMiddleware';
+import AvatarMiddleware from './types/AvatarMiddleware';
+import GroupActivitiesMiddleware from './types/GroupActivitiesMiddleware';
+import ScrollToEndButtonMiddleware from './types/ScrollToEndButtonMiddleware';
 import StyleOptions, { StrictStyleOptions } from './StyleOptions';
+import ToastMiddleware from './types/ToastMiddleware';
+import TypingIndicatorMiddleware from './types/TypingIndicatorMiddleware';
 
 declare const defaultStyleOptions: Required<StyleOptions>;
 
@@ -6,5 +17,18 @@ declare const defaultStyleOptions: Required<StyleOptions>;
 // eslint-disable-next-line no-unused-vars
 declare function normalizeStyleOptions(styleOptions: StyleOptions): StrictStyleOptions;
 
-export type { StrictStyleOptions, StyleOptions };
+export type {
+  ActivityMiddleware,
+  ActivityStatusMiddleware,
+  AttachmentForScreenReaderMiddleware,
+  AttachmentMiddleware,
+  AvatarMiddleware,
+  GroupActivitiesMiddleware,
+  ScrollToEndButtonMiddleware,
+  StrictStyleOptions,
+  StyleOptions,
+  ToastMiddleware,
+  TypingIndicatorMiddleware
+};
+
 export { defaultStyleOptions, normalizeStyleOptions };